+++ crates/miden-lib/src/testing/mock_account_code.rs
@@ -67,15 +67,19 @@ const MOCK_ACCOUNT_CODE: &str = "
 
     # Stack:  [pad(16)]
     # Output: [CODE_COMMITMENT, pad(12)]
-    export.get_code
+    export.get_code_commitment
         exec.account::get_code_commitment
+        # => [CODE_COMMITMENT, pad(16)]
+
+        # truncate the stack
+        swapw dropw
         # => [CODE_COMMITMENT, pad(12)]
     end
 
     # Stack:  [pad(16)]
     # Output: [CODE_COMMITMENT, pad(12)]
-    export.get_storage_commitment
-        exec.account::get_storage_commitment
+    export.compute_storage_commitment
+        exec.account::compute_storage_commitment
         # => [STORAGE_COMMITMENT, pad(16)]
 
         swapw dropw